AvantechPortal/requirements.txt
Django==6.0.4
argon2-cffi>=23.1.0
django-axes>=8.0.0
django-otp>=1.6.0
django-simple-captcha>=0.6.0
python-dotenv>=1.0.0
qrcode>=7.4.2
Pillow>=10.0.0
cryptography>=43.0.0
